+++ src/paging.rs
@@ -1,5 +1,6 @@
 pub struct MaybePager {
     pager: Option<std::process::Child>,
+    abandoned: bool,
 }
 
 impl MaybePager {
@@ -29,7 +30,10 @@ impl MaybePager {
         } else {
             None
         };
-        Self { pager }
+        Self {
+            pager,
+            abandoned: false,
+        }
     }
 
     pub fn wait(&mut self) -> std::io::Result<i32> {
@@ -51,26 +55,40 @@ impl MaybePager {
         }
     }
 
+    /// whether it's ever seen a std::io::ErrorKind::BrokenPipe
+    pub fn is_abandoned(&self) -> bool {
+        self.abandoned
+    }
+
+    /// get the child process's stdin
     fn pipe(&mut self) -> Option<&mut std::process::ChildStdin> {
-        match &mut self.pager {
-            Some(child) => child.stdin.as_mut(),
-            None => None,
-        }
+        self.pager.as_mut().and_then(|child| child.stdin.as_mut())
+    }
+}
+
+pub fn is_broken_pipe<T>(result: &std::io::Result<T>) -> bool {
+    match result {
+        Err(e) => e.kind() == std::io::ErrorKind::BrokenPipe,
+        Ok(_) => false,
     }
 }
 
 impl std::io::Write for MaybePager {
     fn flush(&mut self) -> std::io::Result<()> {
-        match self.pipe() {
+        let result = match self.pipe() {
             Some(pipe) => pipe.flush(),
             None => std::io::stdout().flush(),
-        }
+        };
+        self.abandoned |= is_broken_pipe(&result);
+        result
     }
 
     fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        match self.pipe() {
+        let result = match self.pipe() {
             Some(pipe) => pipe.write(buf),
             None => std::io::stdout().write(buf),
-        }
+        };
+        self.abandoned |= is_broken_pipe(&result);
+        result
     }
 }